Move are_blocked into servicers.blocking
